% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{build_endpoint}
\alias{build_endpoint}
\title{Build an API endpoint URL}
\usage{
build_endpoint(domain, subdomain = NULL, parameters = NULL)
}
\arguments{
\item{domain}{The first component of the API endpoint after base api.municode.com/}

\item{subdomain}{The second component of the API endpoint}

\item{parameters}{Named vector of parameter-value pairs}
}
\value{
Complete API endpoint URL
}
\description{
Build an API endpoint URL
}
\keyword{internal}
