% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{clean_client_columns}
\alias{clean_client_columns}
\title{Standard cleanup for client dataframes}
\usage{
clean_client_columns(df)
}
\arguments{
\item{df}{Client dataframe}
}
\value{
Cleaned dataframe
}
\description{
Standard cleanup for client dataframes
}
\keyword{internal}
