% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{clean_html_content}
\alias{clean_html_content}
\title{Clean HTML content from ordinance text}
\usage{
clean_html_content(content)
}
\arguments{
\item{content}{Character vector containing HTML}
}
\value{
Cleaned character vector
}
\description{
Clean HTML content from ordinance text
}
\keyword{internal}
