% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{convert_html_tables_to_markdown}
\alias{convert_html_tables_to_markdown}
\title{Convert HTML tables within a string to markdown table format}
\usage{
convert_html_tables_to_markdown(html)
}
\arguments{
\item{html}{A character vector of HTML strings containing tables}
}
\value{
A character vector with HTML tables replaced by markdown tables
}
\description{
Convert HTML tables within a string to markdown table format
}
\keyword{internal}
