% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\docType{data}
\name{.municoder_cache}
\alias{.municoder_cache}
\title{Cached versions of API functions for improved performance}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.municoder_cache
}
\description{
Cached versions of API functions for improved performance
}
\keyword{datasets}
