% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_client_content_cached}
\alias{get_client_content_cached}
\title{Get client content with caching}
\usage{
get_client_content_cached(client_id)
}
\arguments{
\item{client_id}{A code corresponding to a given client}
}
\value{
A dataframe with metadata about each product for a client
}
\description{
Cached version of get_client_content(). Results are cached for 6 hours.
}
