% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_client_metadata_cached}
\alias{get_client_metadata_cached}
\title{Get client metadata with caching}
\usage{
get_client_metadata_cached(state_abbreviation, client_name)
}
\arguments{
\item{state_abbreviation}{A two-character state code}

\item{client_name}{The name of a given client}
}
\value{
A dataframe of metadata about the specified client
}
\description{
Cached version of get_client_metadata(). Results are cached for 24 hours.
}
