% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clients.R
\name{get_client_products}
\alias{get_client_products}
\title{Obtain products available for a jurisdiction}
\usage{
get_client_products(client_id)
}
\arguments{
\item{client_id}{A code corresponding to a given client; this can be obtained from \code{get_clients_in_state()}}
}
\value{
A dataframe with metadata about each product for a client (e.g., code of ordinances, zoning, etc.)
}
\description{
Corresponds to, for example: https://api.municode.com/ClientContent/12053
}
\examples{
\dontrun{
get_client_products(980) ## Alexandria, VA
}
}
