% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clients.R
\name{get_clients_in_state}
\alias{get_clients_in_state}
\title{Return metadata for all Municode clients in a given state}
\usage{
get_clients_in_state(state_abbreviation)
}
\arguments{
\item{state_abbreviation}{A two-character state code}
}
\value{
A dataframe of clients
}
\description{
Corresponds to, for example: https://api.municode.com/Clients/stateAbbr?stateAbbr=VA
}
\examples{
\dontrun{
get_clients_in_state("VA")
}
}
