% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_clients_in_state_cached}
\alias{get_clients_in_state_cached}
\title{Get clients in a state with caching}
\usage{
get_clients_in_state_cached(state_abbreviation)
}
\arguments{
\item{state_abbreviation}{A two-character state code}
}
\value{
A dataframe of clients
}
\description{
Cached version of get_clients_in_state(). Results are cached for 24 hours.
}
