% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_codes_content}
\alias{get_codes_content}
\title{Get text and metadata for a given node within a given ordinance
Corresponds to, for example: https://api.municode.com/CodesContent?jobId=426172&nodeId=THZOORALVI&productId=12429}
\usage{
get_codes_content(node_id = NULL, product_id)
}
\arguments{
\item{node_id}{A unique identifier for a node within the specified product (ordinance)}

\item{product_id}{A unique identifier for a product}
}
\value{
A dataframe with the content and metadata pertaining to a specific node within an ordinance
}
\description{
Get text and metadata for a given node within a given ordinance
Corresponds to, for example: https://api.municode.com/CodesContent?jobId=426172&nodeId=THZOORALVI&productId=12429
}
\examples{
get_codes_content(node_id = "SUHITA", product_id = "12429")
}
