% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{get_codes_toc_breadcrumb}
\alias{get_codes_toc_breadcrumb}
\title{Get information about a given node's ancestor(s)}
\usage{
get_codes_toc_breadcrumb(job_id, node_id, product_id)
}
\arguments{
\item{job_id}{A unique identifier for a job}

\item{node_id}{A unique identifier for a node within the specified product (ordinance)}

\item{product_id}{A unique identifier for a product}
}
\description{
Corresponds to, for example: https://api.municode.com/codesToc/breadcrumb?jobId=426172&nodeId=THZOORALVI&productId=12429
}
