% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_codes_toc_children}
\alias{get_codes_toc_children}
\title{Get information about a given node's children
Corresponds to, for example: https://api.municode.com/codesToc/children?jobId=426172&nodeId=ARTIGERE&productId=12429}
\usage{
get_codes_toc_children(job_id, node_id, product_id)
}
\arguments{
\item{job_id}{A unique identifier for a job}

\item{node_id}{A unique identifier for a node within the specified product (ordinance)}

\item{product_id}{A unique identifier for a product}
}
\description{
Get information about a given node's children
Corresponds to, for example: https://api.municode.com/codesToc/children?jobId=426172&nodeId=ARTIGERE&productId=12429
}
\examples{
get_codes_toc_children(job_id = 426172, node_id = "ARTIGERE", product_id = 12429)
}
