% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_core_content_ordinances}
\alias{get_core_content_ordinances}
\title{Return information about all ordinances of a given type (e.g., zoning) for a given year}
\usage{
get_core_content_ordinances(product_id, node_id = NULL)
}
\arguments{
\item{product_id}{A unique code identifying a product (e.g., a zoning ordinance)}

\item{node_id}{The year for which ordinances are requested}
}
\value{
A dataframe comprising ordinances
}
\description{
Corresponds to, for example: https://api.municode.com/CoreContent/Ordinances?nodeId=2023&productId=12429
}
