% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_current_version}
\alias{get_current_version}
\title{Get the current version of a product}
\usage{
get_current_version(product_id)
}
\arguments{
\item{product_id}{A unique identifier for a product}
}
\description{
Corresponds to, for example: https://api.municode.com/Jobs/latest/12429
Returns the current version of a product (e.g., the current zoning ordinance, reflecting amendments)
}
