% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_jobs_product}
\alias{get_jobs_product}
\title{Returns data on all jobs for a given product (e.g., all historical versions of the zoning ordinance)
Corresponds to, for example: https://api.municode.com/Jobs/product/12429}
\usage{
get_jobs_product(product_id)
}
\arguments{
\item{product_id}{A unique identifier for a product}
}
\description{
Returns data on all jobs for a given product (e.g., all historical versions of the zoning ordinance)
Corresponds to, for example: https://api.municode.com/Jobs/product/12429
}
