% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinances.R
\name{get_ordinance_ancestors}
\alias{get_ordinance_ancestors}
\title{Get the ancestors of a given node in an ordinance's Table of Contents}
\usage{
get_ordinance_ancestors(node_id, product_id)
}
\arguments{
\item{node_id}{A unique identifier for a node within the specified product (ordinance)}

\item{product_id}{A unique identifier for a product}
}
\description{
Corresponds to, for example: https://api.municode.com/ordinancesToc/breadcrumb?nodeId=2023&productId=12429
}
