% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinances.R
\name{get_ordinances_toc}
\alias{get_ordinances_toc}
\title{Obtain metadata about a given ordinance over time}
\usage{
get_ordinances_toc(node_id = NULL, product_id)
}
\arguments{
\item{node_id}{A unique identifier for a node within the specified product (ordinance)}

\item{product_id}{A unique identifier for a product}
}
\description{
Corresponds to, for example: https://api.municode.com/ordinancesToc?nodeId=2023&productId=12429
}
\examples{
\dontrun{
get_ordinances_toc(node_id = 2023, product_id = 12429)
}
}
