% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinances.R
\name{get_ordinances_toc_breadcrumb}
\alias{get_ordinances_toc_breadcrumb}
\title{Get the ancestors of a given node in an ordinance's Table of Contents}
\usage{
get_ordinances_toc_breadcrumb(product_id, node_id)
}
\arguments{
\item{product_id}{A unique identifier for a product}

\item{node_id}{A unique identifier for a node within the specified product (ordinance)}
}
\description{
Corresponds to, for example: https://api.municode.com/ordinancesToc/breadcrumb?nodeId=2023&productId=12429
}
