% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_product_metadata}
\alias{get_product_metadata}
\title{Get metadata about a specified product}
\usage{
get_product_metadata(client_id, product_name)
}
\arguments{
\item{client_id}{A unique identifier for a client}

\item{product_name}{The name of a product type, e.g., "Code of Ordinances"}
}
\description{
Corresponds to, for example: https://api.municode.com/Products/name?clientId=12053&productName=code+of+ordinances
}
\examples{
get_product_metadata(client_id = 12053, product_name = "Code of Ordinances")
}
