% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_product_metadata_cached}
\alias{get_product_metadata_cached}
\title{Get product metadata with caching}
\usage{
get_product_metadata_cached(client_id, product_name)
}
\arguments{
\item{client_id}{A unique identifier for a client}

\item{product_name}{The name of a product type}
}
\description{
Cached version of get_product_metadata(). Results are cached for 6 hours.
}
