% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{get_state_by_abbreviation}
\alias{get_state_by_abbreviation}
\title{Look up state metadata given a state abbreviation}
\usage{
get_state_by_abbreviation(state_abbreviation)
}
\arguments{
\item{state_abbreviation}{A two-character state abbreviation}
}
\value{
A dataframe with three columns: \code{state_id}, \code{state_name}, \code{state_abbreviation.}
}
\description{
Corresponds to (for example): https://api.municode.com/States/abbr?stateAbbr=ak
}
