% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_states_abbreviation}
\alias{get_states_abbreviation}
\title{Look up state metadata given a state abbreviation
Corresponds to (for example): https://api.municode.com/States/abbr?stateAbbr=ak}
\usage{
get_states_abbreviation(state_abbreviation)
}
\arguments{
\item{state_abbreviation}{A two-character state abbreviation}
}
\value{
A dataframe with three columns: \code{state_id}, \code{state_name}, \code{state_abbreviation.}
}
\description{
Look up state metadata given a state abbreviation
Corresponds to (for example): https://api.municode.com/States/abbr?stateAbbr=ak
}
