% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_version_history}
\alias{get_version_history}
\title{Version-related API functions
Returns version history for a given product}
\usage{
get_version_history(product_id)
}
\arguments{
\item{product_id}{A unique identifier for a product}
}
\description{
Corresponds to, for example: https://api.municode.com/Jobs/product/12429
Returns all historical versions of a product (e.g., all historical versions of the zoning ordinance)
}
