% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{html_to_markdown}
\alias{html_to_markdown}
\title{Convert HTML content to markdown, preserving table structure}
\usage{
html_to_markdown(html)
}
\arguments{
\item{html}{A character vector of HTML strings}
}
\value{
A character vector of markdown strings
}
\description{
Convert HTML content to markdown, preserving table structure
}
\keyword{internal}
