% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municoder-package.R
\docType{package}
\name{municoder-package}
\alias{municoder}
\alias{municoder-package}
\title{municoder: Interface to the municode.com API}
\description{
Access and programmatically navigate ordinances and associated metadata relating to regulatory documents hosted on municode.com.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wcurrangroome.github.io/municoder/}
}

}
\author{
\strong{Maintainer}: Will Curran-Groome \email{wcurrangroome@gmail.com}

Authors:
\itemize{
  \item Will Curran-Groome \email{wcurrangroome@gmail.com}
}

}
\keyword{internal}
