% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{named_list_to_tibble}
\alias{named_list_to_tibble}
\title{Helper function to convert a named list into a dataframe}
\usage{
named_list_to_tibble(named_list)
}
\arguments{
\item{named_list}{A named list}
}
\description{
Helper function to convert a named list into a dataframe
}
