% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{transform_list_response}
\alias{transform_list_response}
\title{Transform a simple list response to a tidy dataframe}
\usage{
transform_list_response(response, primary_key = "value")
}
\arguments{
\item{response}{Raw API response (list)}

\item{primary_key}{Name of the primary key in the response (default "value")}
}
\value{
A tidy dataframe
}
\description{
Transform a simple list response to a tidy dataframe
}
\keyword{internal}
