% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{transform_nested_client}
\alias{transform_nested_client}
\title{Transform nested Client objects in a dataframe}
\usage{
transform_nested_client(df)
}
\arguments{
\item{df}{Dataframe containing Client columns}
}
\value{
Dataframe with Client columns unnested
}
\description{
Transform nested Client objects in a dataframe
}
\keyword{internal}
