% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{transform_nested_state}
\alias{transform_nested_state}
\title{Transform nested State objects in a dataframe}
\usage{
transform_nested_state(df)
}
\arguments{
\item{df}{Dataframe containing a State column}
}
\value{
Dataframe with State column unnested
}
\description{
Transform nested State objects in a dataframe
}
\keyword{internal}
