% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{build_endpoint}
\alias{build_endpoint}
\title{A helper function to build an API endpoint}
\usage{
build_endpoint(domain, subdomain = NULL, parameters = NULL)
}
\arguments{
\item{domain}{The first component of the API endpoint after base api.municode.com/}

\item{subdomain}{The second component of the API endpoint after api.municode.com/domain}

\item{parameters}{Named vector of parameter-value pairs passed to configure the API endpoint}
}
\description{
A helper function to build an API endpoint
}
