% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{build_endpoint}
\alias{build_endpoint}
\title{API utility functions for municode.com}
\usage{
build_endpoint(domain, subdomain = NULL, parameters = NULL)
}
\arguments{
\item{domain}{The first component of the API endpoint after base api.municode.com/}

\item{subdomain}{The second component of the API endpoint}

\item{parameters}{Named vector of parameter-value pairs}
}
\value{
Complete API endpoint URL
}
\description{
Internal functions for building endpoints and making API calls
}
\keyword{API}
\keyword{Build}
\keyword{URL}
\keyword{an}
\keyword{endpoint}
\keyword{internal}
