% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clients.R
\name{get_client_metadata}
\alias{get_client_metadata}
\title{Client-related API functions
Returns metadata about a given client}
\usage{
get_client_metadata(state_abbreviation, client_name)
}
\arguments{
\item{state_abbreviation}{A two-character state code}

\item{client_name}{The name of a given client}
}
\value{
A dataframe of metadata about the specified client
}
\description{
Corresponds to, for example: https://api.municode.com/Clients/name?stateAbbr=VA&clientName=Alexandria
}
\examples{
\dontrun{
get_client_metadata("VA", "Alexandria")
}
}
