% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_endpoint}
\alias{get_endpoint}
\title{A helper function to send a GET request to an API endpoint}
\usage{
get_endpoint(endpoint)
}
\arguments{
\item{endpoint}{An API endpoint}
}
\description{
A helper function to send a GET request to an API endpoint
}
