% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_jobs_latest}
\alias{get_jobs_latest}
\title{Get data on the most recent job}
\usage{
get_jobs_latest(product_id)
}
\arguments{
\item{product_id}{A unique identifier for a product}
}
\description{
Corresponds to, for example: https://api.municode.com/Jobs/latest/12429
Returns the current version of a product (e.g., the current zoning ordinance, reflecting amendments)
}
