% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_jobs_latest}
\alias{get_jobs_latest}
\title{This returns data on the most recent job (e.g., the current zoning ordinance, reflecting amendments)}
\usage{
get_jobs_latest(product_id)
}
\arguments{
\item{product_id}{A unique identifier for a product}
}
\description{
Corresponds to, for example: https://api.municode.com/Jobs/latest/12429
}
