% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_ordinances_toc}
\alias{get_ordinances_toc}
\title{Obtain metadata about a given ordinance over time
Corresponds to, for example: https://api.municode.com/ordinancesToc?nodeId=2023&productId=12429}
\usage{
get_ordinances_toc(product_id, node_id = NA)
}
\arguments{
\item{product_id}{A unique identifier for a product}

\item{node_id}{A unique identifier for a node within the specified product (ordinance)}
}
\description{
Obtain metadata about a given ordinance over time
Corresponds to, for example: https://api.municode.com/ordinancesToc?nodeId=2023&productId=12429
}
\examples{
get_ordinances_toc(node_id = 2023, product_id = 12429)
}
