% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municode_endpoints.R
\name{get_states}
\alias{get_states}
\title{Return metadata about states (and similar) as used by municode
Corresponds to: https://api.municode.com/States/}
\usage{
get_states()
}
\value{
A dataframe with three columns: \code{state_id}, \code{state_name}, \code{state_abbreviation.}
}
\description{
Return metadata about states (and similar) as used by municode
Corresponds to: https://api.municode.com/States/
}
