% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinances.R
\name{list_ordinances}
\alias{list_ordinances}
\title{Ordinance-related API functions (historical ordinance versions)
Return information about all ordinances of a given type for a given year}
\usage{
list_ordinances(product_id, node_id = NULL)
}
\arguments{
\item{product_id}{A unique code identifying a product (e.g., a zoning ordinance)}

\item{node_id}{The year for which ordinances are requested}
}
\value{
A dataframe comprising ordinances
}
\description{
Corresponds to, for example: https://api.municode.com/CoreContent/Ordinances?nodeId=2023&productId=12429
}
