% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{transform_list_response}
\alias{transform_list_response}
\title{Common data transformation helpers for municode API responses}
\usage{
transform_list_response(response, primary_key = "value")
}
\arguments{
\item{response}{Raw API response (list)}

\item{primary_key}{Name of the primary key in the response (default "value")}
}
\value{
A tidy dataframe
}
\description{
Internal functions to standardize data processing across the package
}
\keyword{Transform}
\keyword{a}
\keyword{dataframe}
\keyword{internal}
\keyword{list}
\keyword{response}
\keyword{simple}
\keyword{tidy}
\keyword{to}
